experimental refactor + fused kernels
