+++ examples/CMakeLists.txt
@@ -6,6 +6,8 @@
 #   cmake --build . --target example_etth example_ptbxl
 set(NOTT_EXAMPLES
     example_overhead      analysis/overhead.cpp
+    example_bench_attention analysis/bench_attention.cpp
+    example_bench_memory_arena analysis/bench_memory_arena.cpp
     example_cifar10       classification/images/cifar10.cpp
     example_dubai_segment classification/masks/dubai_segment.cpp
     example_ptbxl         classification/timeseries/ptbxl.cpp
