+++ src/data/transform/dimensionality_reduction/rpca.hpp
@@ -12,7 +12,7 @@
 #include <torch/torch.h>
 
 #include "common.hpp"
-#include "../../../core.hpp"
+#include "../../../core/core.hpp"
 
 namespace Nott::Data::Transform::DimensionalityReduction {
     template <bool BufferVRAM = false, class DevicePolicyT = Core::DevicePolicy<BufferVRAM>>
