+++ src/plot/details/reliability/reliability_curve_utils.hpp
@@ -12,7 +12,7 @@
 
 #include <torch/torch.h>
 
-#include "../../../core.hpp"
+#include "../../../core/core.hpp"
 
 namespace Nott::Plot::Details::Reliability::Curves {
     struct BinarySeries {
