+++ tests/CMakeLists.txt
@@ -9,9 +9,11 @@ add_executable(nott_tests EXCLUDE_FROM_ALL
     initialization_test.cpp
     regularization_test.cpp
     optimizer_test.cpp
+    attention_kernel_test.cpp
     layer_test.cpp
     block_test.cpp
     links_test.cpp
+    model_io_test.cpp
     metric_ground_truth_test.cpp
     training_integration_test.cpp
 )
